requirements.txt
boto3==1.9.188
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-extensions==2.1.4
django-filter==2.1.0
django-filters==0.2.1
django-model-utils==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docutils==0.14
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.5.0
Markdown==3.0.1
MarkupSafe==1.1.1
mysqlclient==1.3.13
openpyxl==2.6.2
pycrypto==2.6.1
Pygments==2.3.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
